+++ src/commands/index.ts
@@ -1,3 +1,5 @@
 export * from './helloWorld';
 export * from './installUpdate';
 export * from './pickVersion';
+export * from './pinVersion';
+export * from './switchDepType';