+++ src/providers/index.ts
@@ -1,5 +1,6 @@
 export * from './PackagesProvider';
 export * from './PackageItem';
+export * from './PackageDetailItem';
 export * from './GroupItem';
 export * from './WorkspaceFolderItem';
 export * from './FilterBarItem';